+++ app/layout.tsx
@@ -1,3 +1,5 @@
+import AuthContext from './AuthContext';
+
 export default function RootLayout({
   children,
 }: {
@@ -6,7 +8,11 @@ export default function RootLayout({
   return (
     <html>
       <head />
-      <body>{children}</body>
+      <body>
+        <AuthContext>
+          {children}
+        </AuthContext>
+      </body>
     </html>
   )
 }
