--- app/pages/api/test.ts
@@ -1,5 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next'
-
-export default function handler(req: NextApiRequest, res: NextApiResponse) {
-  res.status(200).json({ name: 'poo poo pee pee' })
-}
