--- pages/api/auth.ts
@@ -1,24 +1,0 @@
-//all dynamic routes for NextAuth
-//contains all gloabal NextAuth configs
-
-import NextAuth from 'next-auth';
-import GithubProvider from 'next-auth/providers/github'
-import GoogleProvider from 'next-auth/providers/google'
-
-
-export const authOptions = {
-    // Configure one or more authentication providers
-    providers: [
-      GithubProvider({
-        clientId: process.env.GITHUB_ID,
-        clientSecret: process.env.GITHUB_SECRET,
-      }),
-      // ...add more providers here
-      GoogleProvider({
-        clientId: process.env.GOOGLE_ID,
-        clientSecret: process.env.GOOGLE_SECRET
-    }),
-    ],
-  }
-  
-  export default NextAuth(authOptions)
